\name{createH5Attribute}
\alias{createH5Attribute}
\title{ fille me
}
\description{
fill me
}
\usage{
createH5Attribute(h5Obj, attrName, attrValue, ...)
}
\arguments{
  \item{h5Obj}{
  Oject in which attribute is located.
}
  \item{attrName}{
    Name of attribute
}
  \item{attrValue}{
  Actual data values
}
  \item{\dots}{
    extra arguments
}
}
\details{
  fill me
}
\value{
  return values of the function ... 
}
\seealso{
  readH5Attribute
}
\examples{
2 + 2
}
\keyword{dataset}

