\name{createH5Group}
\alias{createH5Group}
\title{
   Fill in
}
\description{
  Fill in
}
\usage{
createH5Group(h5Obj, groupName, ...)
}
\arguments{
  \item{h5Obj}{
  Fill in
}
  \item{groupName}{
  Describe this
}
  \item{\dots}{
  Describe this
}
}
\details{
 more here
}
\value{
  Describe the value returned
  If it is a LIST, use
  \item{comp1 }{Description of 'comp1'}
  \item{comp2 }{Description of 'comp2'}
}
\references{
  Blah 
}
\note{
  Further notes
}
\seealso{
  Helpful link
}
\examples{
  Example here
}
\keyword{key}

